添加测试APP
